Testing out styling of Atom notifications.
I don't really like the way Atom's popover notification cards are
styled; the colors don't really work well and they don't fit with
the rest of the UI. I therefore implemented a quick style for the
error notifications to see if it could be done any better.

So far I'm liking the results better than the way things originally
were, but I'm not done with these yet. I also need to implement
styles for the remaining notification types.